Move LinkedIn cron to Job (#269)
* create pr

* Potential fix for code scanning alert no. 238: Incomplete string escaping or encoding


* fix feedback

* Use serverless job architecture

* Potential fix for code scanning alert no. 240: Workflow does not contain permissions


* Add URL in return data

---------
