+++ .github/workflows/main-pipeline.yml
@@ -1,5 +1,9 @@
 name: Main Pipeline
 
+permissions:
+  contents: read
+  pull-requests: write
+
 on:
   workflow_run:
     workflows: ["Tests"]
