+++ api/package.json
@@ -29,6 +29,7 @@
     "dotenv": "16.5.0",
     "express": "4.21.0",
     "express-rate-limit": "^7.5.0",
+    "fast-xml-parser": "^5.2.5",
     "geoip-lite": "1.4.10",
     "helmet": "8.1.0",
     "isbot": "5.1.28",
