+++ api/src/jobs/linkedin/types.ts
@@ -1,0 +1,38 @@
+export interface LinkedInJob {
+  partnerJobId: string;
+  company: string;
+  title: string;
+  description: string;
+  applyUrl: string;
+  companyId: string;
+  location: string;
+  lastBuildDate?: string;
+  publisherURL?: string;
+  publisher?: string;
+  expectedJobCount?: number;
+  alternateLocations?: string[];
+  city?: string;
+  state?: string;
+  country?: string;
+  postalCode?: string;
+  isRemote?: "On-site" | "Hybrid" | "Remote";
+  workplaceTypes?: string;
+  industryCodes?: {
+    industryCode: string;
+  }[];
+  industryCode?: string;
+  industry?: string; // Is it really used
+  experienceLevel?: string;
+  jobFunctions?: string[];
+  jobFunction?: string;
+  jobtype?: string;
+  Skills?: string[];
+  salaries?: string[];
+  highEnd?: number;
+  lowEnd?: number;
+  currencyCode?: string;
+  period?: string;
+  type?: string;
+  listDate?: string;
+  expirationDate?: string;
+}
